% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age.R
\docType{data}
\name{age}
\alias{age}
\title{Age effect on body temperature}
\format{
A data frame with 61 rows and 99 columns. Each row stands for an age (from 20 to 80) and each column stands for a quantile level (from 1\% to 99\%). The number corresponding to age \code{X} and quantile \code{Y} denotes the deviation from the baseline (stored in the dataset _baseline_) of a subject with age \code{X} in the \code{Y} quantile of his or her body temperature.

#' @source the dataset of age effect is extracted from an additive quantile regression model of body temperature.
}
\usage{
age
}
\description{
A dataset describing how age affects the distribution (quantiles) of body temperature.
}
\keyword{datasets}
