% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function.R
\name{celsius_to_kelvin}
\alias{celsius_to_kelvin}
\title{Celsius to Kelvin conversion}
\usage{
celsius_to_kelvin(temp)
}
\arguments{
\item{temp}{A numeric variable. A temperature expressed in Celsius}
}
\description{
This function takes a temperature expressed in Celsius and convert it to Kelvin.
}
\examples{
celsius_to_kelvin(0)

}
