% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial.R
\docType{data}
\name{initial}
\alias{initial}
\title{Initial input for the visualization app of body temperature}
\format{
A data frame with 1 row and 8 covariates: gender, age, sizeUnit, height, height2, weight, tempUnit (temperature unit) and temp (temperature).
}
\usage{
initial
}
\description{
Before users input their covariates, the visualization is carried out based on the initial input for demonstration.
}
\keyword{datasets}
