% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function.R
\name{plot_trend}
\alias{plot_trend}
\title{Plots temperatues of every month?}
\usage{
plot_trend(dat)
}
\arguments{
\item{dat}{Input data in wide format}
}
\description{
Convert wide to long temperature datasets
}
\examples{
data(tuggeranong)
plot_trend(tuggeranong)

}
