% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_app.R
\name{temp_app}
\alias{temp_app}
\title{temp_app}
\usage{
temp_app()
}
\description{
This function calls the visualization app. The app takes in the covariates of a subject (gender, age, weight, height, temperature) and outputs the 1\%, 5\%, 25\%, 50\%, 75\%, 95\%, 99\% percentiles of the body temperature from 7:00 to 18:00.
}
