% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temperature.to.percentile.R
\name{temperature.to.percentile}
\alias{temperature.to.percentile}
\title{temperature.to.percentile}
\usage{
temperature.to.percentile(data)
}
\arguments{
\item{data}{a data.frame recording covariates of subjects, and each row represents a subject. The \code{data} should contain at least six named columns:
\itemize{
\item temperature: temperature of the subject (in Fahrenheit). The \code{temperature} should be numeric.
\item time: the time when the temperature is taken. The \code{time} should be a character of the format hh:mm, e.g., "08:30" and "12:00". The \code{time} should be no earlier than 07:00 or later than 18:00 within a day.
\item gender: gender of the subject. The \code{gender} should be a character. There are three possible values: "male", "female.pre" for ladies before menopause, and "female.post" for ladies after menopause. If the menopause condition in not available, a default rule regards female no younger than 40 as post-menopause.
\item age: age of the subject. The \code{age} should be numeric and will be rounded to the nearest integer. The \code{age} should be no smaller than 20 or larger than 80.
\item height: height of the subject (in meter). The \code{height} should be numeric and will be rounded to two decimal places. The \code{height} should be no smaller than 1.38 or larger than 2.13.
\item weight: weight of the subject (in kilogram). The \code{weight} should be numeric and will be rounded to the nearest integer. The \code{weight} should be no smaller than 30 or larger than 181.
}}
}
\value{
The function returns of a vector representing the percentiles corresponding to the given temperatures. We use 101 or -1 to denote a temperature that is above the 99% percentile or below the 1% percentile.
}
\description{
This function computes the percentile corresponding to a body temperature for a given subject.
}
\examples{
temperature.to.percentile(data)

}
