% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function.R
\name{Wide_To_Long_Temperature}
\alias{Wide_To_Long_Temperature}
\title{Convert wide to long temperature datasets}
\usage{
Wide_To_Long_Temperature(
  dat,
  id.vars = "Year",
  variable.name = "Month",
  value.name = "Temperature"
)
}
\arguments{
\item{dat}{x}

\item{id.vars}{x}

\item{variable.name}{x}

\item{value.name}{x}
}
\description{
Convert wide to long temperature datasets
}
\examples{
data(tuggeranong)
Wide_To_Long_Temperature(tuggeranong)

}
