% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function.R
\name{kelvin_to_celsius}
\alias{kelvin_to_celsius}
\title{Kelvin to Celsius conversion}
\usage{
kelvin_to_celsius(temp)
}
\arguments{
\item{temp}{A numeric variable. A temperature expressed in Kelvin}
}
\description{
This function takes a temperature expressed in Kelvin and convert it to Celsius.
}
\examples{
kelvin_to_celsius(273.15)

}
